+++ packages/manager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sandagent/manager
 
+## 0.2.3
+
+### Patch Changes
+
+- refactor runner
+
 ## 0.2.2
 
 ### Patch Changes