+++ packages/sandbox-sandock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sandagent/sandbox-sandock
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @sandagent/manager@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes