+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sandagent/sdk
 
+## 0.2.3
+
+### Patch Changes
+
+- refactor runner
+- Updated dependencies
+  - @sandagent/manager@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes